Connector now supports optional getPasswordPolicies iterator. There is no requirement that this be implemented, nor any concrete requirement on how the connector reports users that belong to these objects.
In the case of AD we report the PSOs and then we can establish user/group membership when iterating over the users by looking at the msDS-PSOApplied attributes of the user/group object.

This shoud allow us to reconcile PSO and only add a small operation to each user/group reconcile to link their password policies with their accounts.