Re-introduced some of the behavior of NssTokenDatabase in 'DefaultNssTokenDatabase'. This holds a static instance that will be used by DefaultEncoderManager if it has been set. HS continues to use NssTokenDatabase directly so it will not be affected by this.(and it can continue to manually register the encoders).
Both NAM and VMCentre will now use DefaultTokenDatabase so will act as they previously did.

I believe this whole problem stems from the Connectors use of EncoderManager. They need one, and there is currently no mechanism for setting one. These in turn need an NssTokenDatabase, and in NAM's case that currently needs to be shared throughout the system. The long term fix is to change I4J so that an encoder manager (and maybe individual encoders) instance may be set on a connector instance, allowing NssTokenDatabase and EncoderManager to lose their static fields (get rid of the 'Default') implementation and leave it entirely up to the app to configure it correctly.